[ENHANCEMENT] Extract MimeWalk in a separate class
